Satellite death sync, lost worker recovery, translucent HUD, tint fix
- TOPO_DEATH_SYNC: satellite reports worker deaths to queen for registry
- Lost worker recovery: queen tracks remote workers, respawns after 30s
  satellite disconnect
- Persistence gap: create IdentityRecord for any worker missing one
  (fixes conkers vanishing on reboot after cross-module handoff)
- Queen position persisted correctly on Case B migration (flush_manifest)
- Founding egg gate: check existing brood unconditionally, fix
  total_duration_ms not being saved to BroodRecord
- Translucent HUD: alpha-blend over framebuffer using palette bg_a,
  rotation-aware native buffer addressing
- Tint system: full seed range 1-255 for colour, separate age_grey
  parameter for desaturation (fixes ~half of conkers having no tint)
- API: tint_seed + scale_factor added to list and detail endpoints
- App: ConkerSprite tint algorithm updated to match firmware v70

+++ firmware/src/persistence.cpp
@@ -276,7 +276,7 @@ bool ConkerRegistry::_write_brood(const BroodRecord& rec) {
     doc["hunger"]           = rec.hunger;
     doc["food_invested"]    = rec.food_invested;
     doc["stage_elapsed_ms"] = stage_elapsed;
-    doc["total_duration_ms"] = rec.total_duration_ms;
+    doc["total_duration_ms"] = (unsigned long)rec.total_duration_ms;
     doc["born_unix"]        = rec.born_unix;
 
     char buf[256];
@@ -396,8 +396,8 @@ bool ConkerRegistry::_load_brood_records() {
                         b.y              = doc["y"] | 0;
                         b.hunger         = doc["hunger"] | 0.0f;
                         b.food_invested  = doc["food_invested"] | 0.0f;
-                        b.total_duration_ms = doc["total_duration_ms"] | 0;
-                        b.born_unix      = doc["born_unix"] | 0;
+                        b.total_duration_ms = doc["total_duration_ms"] | (uint32_t)0;
+                        b.born_unix      = doc["born_unix"] | (uint32_t)0;
                         // Convert elapsed time back to absolute millis
                         uint32_t elapsed = doc["stage_elapsed_ms"] | 0;
                         b.stage_start_ms = millis() - elapsed;
